refactor: 폴더 구조 변경
- jwt 폴더 추가
